Math BS and fix bmp parser

+++ src/libm/mat4_algebra.c
@@ -6,7 +6,7 @@
 /*   By: tlernoul <tlernoul@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/23 15:52:10 by tlernoul          #+#    #+#             */
-/*   Updated: 2020/06/23 16:08:15 by tlernoul         ###   ########.fr       */
+/*   Updated: 2020/06/24 15:42:28 by tlernoul         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,27 +31,27 @@ t_mat4 		m4_transpose(t_mat4 m)
 
 static void	ret_axis_x_rot_matrix(t_mat4 *m, float theta)
 {
-	m->m[5] = cos(theta);
-	m->m[6] = sin(theta);
-	m->m[9] = -sin(theta);
-	m->m[10] = cos(theta);
+	m->m[5] = cosf(theta);
+	m->m[6] = sinf(theta);
+	m->m[9] = -sinf(theta);
+	m->m[10] = cosf(theta);
 }
 
 static void	ret_axis_y_rot_matrix(t_mat4 *m, float theta)
 {
-	m->m[0] = cos(theta);
-	m->m[2] = -sin(theta);
-	m->m[8] = sin(theta);
-	m->m[10] = cos(theta);
+	m->m[0] = cosf(theta);
+	m->m[2] = -sinf(theta);
+	m->m[8] = sinf(theta);
+	m->m[10] = cosf(theta);
 }
 
 
 static void	ret_axis_z_rot_matrix(t_mat4 *m, float theta)
 {
-	m->m[0] = cos(theta);
-	m->m[1] = sin(theta);
-	m->m[4] = -sin(theta);
-	m->m[5] = cos(theta);
+	m->m[0] = cosf(theta);
+	m->m[1] = sinf(theta);
+	m->m[4] = -sinf(theta);
+	m->m[5] = cosf(theta);
 }
 
 t_mat4 		m4_rotate_axis(t_mat4 m, t_axis axis, float angle)
@@ -60,7 +60,7 @@ t_mat4 		m4_rotate_axis(t_mat4 m, t_axis axis, float angle)
 	float	theta;
 
 	m4_set(&r, (float)0x7FFFFFFF);
-	theta = angle * (M_PI / 180);
+	theta = angle * (float)(M_PI / 180);
 	if (axis == AXIS_X)
 	{
 		ret_axis_x_rot_matrix(&r, theta);
